# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Law OA Go API
  description: |
    Law OA Go 提供现代化的 RESTful API，基于 Go 1.23+ 标准构建，采用统一响应格式、类型安全处理和完善的中间件体系。
    
    ## 核心特性
    - ✅ **统一响应格式**: 所有API使用一致的响应结构
    - ✅ **类型安全**: 基于 Go generics 的类型安全处理
    - ✅ **现代中间件**: CORS、限流、JWT认证、超时控制
    - ✅ **完善错误处理**: 标准化错误响应和建议信息
    - ✅ **请求追踪**: 每个请求都有唯一ID用于调试
    - ✅ **分页支持**: 统一的分页参数和响应格式
    
    ## 基础信息
    - **Base URL**: http://localhost:8080/api/v1
    - **认证方式**: Bearer Token (JWT)
    - **数据格式**: JSON
    - **API版本**: v1
    - **Go版本**: 1.23+
  version: 1.0.0
  contact:
    name: Law OA Go Team
    email: support@lawfirm.com
    url: https://github.com/law-oa-go
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 本地开发服务器
  - url: https://api.lawfirm.com/v1
    description: 生产服务器
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT认证令牌，通过登录接口获取
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        meta:
          $ref: '#/components/schemas/Meta'
      required:
        - success
        - meta

    # 成功响应
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              description: 响应数据
          required:
            - data

    # 分页响应
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/ErrorDetail'
          required:
            - error

    # 错误详情
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          description: 错误码
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: 错误消息
          example: "Validation failed"
        details:
          type: string
          description: 错误详情
          example: "Email format is invalid"
        suggestions:
          type: array
          description: 建议解决方案
          items:
            type: string
          example: ["Check email format"]
      required:
        - code
        - message

    # 元数据
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
          example: "2024-01-01T00:00:00Z"
        request_id:
          type: string
          description: 请求ID
          example: "req_123456"
        version:
          type: string
          description: API版本
          example: "1.0.0"
      required:
        - timestamp
        - request_id
        - version

    # 分页信息
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: 用户ID
          example: 1
        name:
          type: string
          description: 用户姓名
          example: "系统管理员"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "admin@lawfirm.com"
        role:
          type: string
          description: 用户角色
          enum: [admin, user, lawyer]
          example: "admin"
        phone:
          type: string
          description: 电话号码
          example: "13800138000"
        avatar:
          type: string
          description: 头像URL
          example: "https://example.com/avatar.jpg"
        status:
          type: string
          description: 用户状态
          enum: [active, inactive, suspended]
          example: "active"
        created_at:
          type: string
          format: date-time
          description: 创建时间
          example: "2024-01-01T00:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: 更新时间
          example: "2024-01-01T00:00:00Z"
      required:
        - id
        - name
        - email
        - role
        - status
        - created_at
        - updated_at

    # 客户模型
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "admin@lawfirm.com"
        password:
          type: string
          format: password
          description: 密码
          example: "password123"
      required:
        - email
        - password

    # 注册请求
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          description: 用户姓名
          example: "新用户"
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "newuser@example.com"
        password:
          type: string
          format: password
          description: 密码
          example: "password123"
        phone:
          type: string
          description: 电话号码
          example: "13800138000"
        role:
          type: string
          description: 用户角色
          enum: [admin, user, lawyer]
          example: "user"
      required:
        - name
        - email
        - password
        - phone
        - role

    # 创建客户请求
...
paths:
...
  /auth/login:
    post:
      tags:
        - 认证
      summary: 用户登录
      description: 使用邮箱和密码进行登录认证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            description: JWT令牌
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          expires_at:
                            type: string
                            format: date-time
                            description: 令牌过期时间
                            example: "2024-12-31T23:59:59Z"
                          user:
                            $ref: '#/components/schemas/User'
                        required:
                          - token
                          - expires_at
                          - user
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/register:
    post:
      tags:
        - 认证
      summary: 用户注册
      description: 创建新用户账户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            description: JWT令牌
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          expires_at:
                            type: string
                            format: date-time
                            description: 令牌过期时间
                            example: "2024-12-31T23:59:59Z"
                          user:
                            $ref: '#/components/schemas/User'
                        required:
                          - token
                          - expires_at
                          - user
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 用户已存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - 认证
      summary: 刷新令牌
      description: 使用现有令牌获取新的访问令牌
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: 当前JWT令牌
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              required:
                - token
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          token:
                            type: string
                            description: 新的JWT令牌
                            example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                          expires_at:
                            type: string
                            format: date-time
                            description: 令牌过期时间
                            example: "2024-12-31T23:59:59Z"
                        required:
                          - token
                          - expires_at
        '401':
          description: 令牌无效或已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 用户管理接口
...